Completed first draft of Stable Matching algorithm.
The project still needs comments as well as a way to read function to
fill array's and a write array to write the stable matching algorithm
resualts.

What you should do is read a document then fill the array's then write a
file with an easy to read matching style.